+++ Cargo.toml
@@ -1,17 +1,18 @@
+[workspace]
+members = ["core", "ibus"]
+default-members = ["core", "ibus"]
+
 [package]
 description = "Bộ gõ tiếng Việt mã nguồn mở đa hệ điều hành Gõ Key"
 edition = "2021"
 name = "goxkey"
 version = "0.3.1"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 env_logger = "0.10.0"
 libc = "0.2.139"
 log = "0.4.17"
 vi = "0.6.2"
-bitflags = "1.3.2"
 druid = { features = [
     "image",
     "png",
@@ -19,6 +20,7 @@ druid = { features = [
 once_cell = "1.17.0"
 auto-launch = "0.5.0"
 nom = "7.1.3"
+goxkey-core = { path = "core", features = ["druid"] }
 
 [target.'cfg(target_os="macos")'.dependencies]
 core-foundation = "0.9.3"
