+++ Makefile
@@ -28,3 +28,65 @@ update-cask:
 	sed -i '' 's/version ".*"/version "$(VERSION)"/' Casks/goxkey.rb
 	sed -i '' 's/sha256 ".*"/sha256 "$(SHA256)"/' Casks/goxkey.rb
 	@echo "Casks/goxkey.rb updated → version=$(VERSION) sha256=$(SHA256)"
+
+# ──────────────────────────────────────────────
+# IBus engine (Linux)
+# ──────────────────────────────────────────────
+
+IBUS_COMPONENT_DIR ?= /usr/share/ibus/component
+IBUS_BIN_DIR      ?= $(HOME)/.local/bin
+
+.PHONY: ibus-setup ibus-uninstall ibus-reinstall
+
+ibus-setup: ibus-build ibus-install-xml ibus-restart
+
+ibus-reinstall: ibus-build ibus-install ibus-install-xml ibus-restart
+
+ibus-build:
+	cargo build --release -p goxkey-ibus
+	install -d $(IBUS_BIN_DIR)
+	install -m 755 target/release/goxkey-ibus $(IBUS_BIN_DIR)/goxkey-ibus
+
+ibus-install:
+	install -d $(IBUS_BIN_DIR)
+	install -m 755 target/release/goxkey-ibus $(IBUS_BIN_DIR)/goxkey-ibus
+
+ibus-install-xml:
+	sudo install -d $(IBUS_COMPONENT_DIR)
+	{ \
+	  echo '<?xml version="1.0" encoding="utf-8"?>'; \
+	  echo '<component>'; \
+	  echo '  <name>org.freedesktop.IBus.Goxkey</name>'; \
+	  echo '  <description>Goxkey Component</description>'; \
+	  echo "  <exec>$(IBUS_BIN_DIR)/goxkey-ibus</exec>"; \
+	  echo '  <version>$(VERSION)</version>'; \
+	  echo '  <author>Huy Tran</author>'; \
+	  echo '  <license>MIT</license>'; \
+	  echo '  <homepage>https://github.com/huytd/goxkey</homepage>'; \
+	  echo '  <textdomain>goxkey</textdomain>'; \
+	  echo '  <engines>'; \
+	  echo '    <engine>'; \
+	  echo '      <name>goxkey</name>'; \
+	  echo '      <language>vi</language>'; \
+	  echo '      <license>MIT</license>'; \
+	  echo '      <author>Huy Tran</author>'; \
+	  echo '      <layout>us</layout>'; \
+	  echo '      <longname>GoKey Vietnamese</longname>'; \
+	  echo '      <description>Vietnamese Input Method (Telex, VNI, Telex+VNI)</description>'; \
+	  echo '      <rank>0</rank>'; \
+	  echo '    </engine>'; \
+	  echo '  </engines>'; \
+	  echo '</component>'; \
+	} | sudo tee $(IBUS_COMPONENT_DIR)/goxkey.xml > /dev/null
+	@echo "goxkey.xml installed to $(IBUS_COMPONENT_DIR)"
+
+ibus-restart:
+	@echo "Restarting IBus..."
+	ibus restart 2>/dev/null || (ibus-daemon --daemonize 2>/dev/null) || true
+	@echo "Done. Select Gõ Key in your input method settings."
+
+ibus-uninstall:
+	rm -f $(IBUS_COMPONENT_DIR)/goxkey.xml
+	rm -f $(IBUS_BIN_DIR)/goxkey-ibus
+	ibus restart 2>/dev/null || true
+	@echo "Goxkey IBus engine uninstalled."
