+++ core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "goxkey-core"
+version = "0.3.1"
+edition = "2021"
+
+[dependencies]
+vi = "0.6.2"
+bitflags = "1.3.2"
+once_cell = "1.17.0"
+log = "0.4.17"
+nom = "7.1.3"
+druid = { git = "https://github.com/huytd/druid", branch = "master", optional = true }
+
+[features]
+default = []
