+++ core/src/input.rs
@@ -1,24 +1,16 @@
 use std::collections::BTreeMap;
-use std::{collections::HashMap, fmt::Display, str::FromStr};
+use std::fmt::Display;
+use std::str::FromStr;
 
-use druid::{Data, Target};
 use log::debug;
-use once_cell::sync::{Lazy, OnceCell};
-use rdev::{Keyboard, KeyboardState};
+use once_cell::sync::Lazy;
 use vi::TransformResult;
 
-use crate::platform::{get_active_app_name, KeyModifier};
-use crate::{
-    config::CONFIG_MANAGER, hotkey::Hotkey, platform::is_in_text_selection, ui::UPDATE_UI,
-    UI_EVENT_SINK,
-};
+use crate::config::CONFIG_MANAGER;
+use crate::hotkey::Hotkey;
+use crate::key_modifier::KeyModifier;
 
-// According to Google search, the longest possible Vietnamese word
-// is "nghiêng", which is 7 letters long. Add a little buffer for
-// tone and marks, I guess the longest possible buffer length would
-// be around 10 to 12.
 const MAX_POSSIBLE_WORD_LENGTH: usize = 10;
-const MAX_DUPLICATE_LENGTH: usize = 4;
 const TONE_DUPLICATE_PATTERNS: [&str; 17] = [
     "ss", "ff", "jj", "rr", "xx", "ww", "kk", "tt", "nn", "mm", "yy", "hh", "ii", "aaa", "eee",
     "ooo", "ddd",
@@ -29,17 +21,8 @@ pub static mut HOTKEY_MODIFIERS: KeyModifier = KeyModifier::MODIFIER_NONE;
 pub static mut HOTKEY_MATCHING: bool = false;
 pub static mut HOTKEY_MATCHING_CIRCUIT_BREAK: bool = false;
 
-pub const PREDEFINED_CHARS: [char; 47] = [
-    'a', '`', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=', 'q', 'w', 'e', 'r', 't',
-    'y', 'u', 'i', 'o', 'p', '[', ']', 's', 'd', 'f', 'g', 'h', 'j', 'k', 'l', ';', '\'', '\\',
-    'z', 'x', 'c', 'v', 'b', 'n', 'm', ',', '.', '/',
-];
-
 pub const STOP_TRACKING_WORDS: [&str; 4] = [";", "'", "?", "/"];
-/// In w-literal mode, replace standalone 'w' with placeholder bytes that the telex
-/// engine ignores (falls through to `_ => Transformation::Ignored`), then restore them
-/// after transformation. A 'w' is "standalone" when NOT preceded by a Horn/Breve-eligible
-/// vowel — those cases (uw→ư, ow→ơ, aw→ă) should still be handled by telex normally.
+
 enum CapPattern {
     Lower,
     TitleCase,
@@ -75,8 +58,6 @@ fn apply_cap_pattern(s: &str, pattern: CapPattern) -> String {
 }
 
 fn mask_standalone_w(buffer: &str) -> String {
-    // Characters that can accept Horn (w) modification: u, o and all their toned forms.
-    // Characters that can accept Breve (w) modification: a and all its toned forms.
     const HORN_BREVE_ELIGIBLE: &str = "uoaUOA\u{01b0}\u{01a1}\u{0103}\
          \u{00fa}\u{00f3}\u{00e1}\u{00f9}\u{00f2}\u{00e0}\
          \u{1ee7}\u{1ecf}\u{1ea3}\u{0169}\u{00f5}\u{00e3}\u{1ecd}\u{1ea1}\
@@ -87,17 +68,12 @@ fn mask_standalone_w(buffer: &str) -> String {
     for (i, &ch) in chars.iter().enumerate() {
         if ch == 'w' || ch == 'W' {
             let preceded_by_eligible = i > 0 && HORN_BREVE_ELIGIBLE.contains(chars[i - 1]);
-            // Also pass through when this 'w' follows a 'w' that was itself
-            // preceded by an eligible vowel (e.g. "aww", "uww", "oww").
-            // This lets telex see the full "ww" sequence and undo the
-            // Horn/Breve modification, producing the raw text.
             let preceded_by_w_after_eligible = i >= 2
                 && (chars[i - 1] == 'w' || chars[i - 1] == 'W')
                 && HORN_BREVE_ELIGIBLE.contains(chars[i - 2]);
             if preceded_by_eligible || preceded_by_w_after_eligible {
-                result.push(ch); // let telex transform it: uw→ư, ow→ơ, aw→ă, or ww→undo
+                result.push(ch);
             } else {
-                // Mask it — telex ignores \x01/\x02, we restore them after transform
                 result.push(if ch == 'w' { '\x01' } else { '\x02' });
             }
         } else {
@@ -107,92 +83,33 @@ fn mask_standalone_w(buffer: &str) -> String {
     result
 }
 
-pub fn get_key_from_char(c: char) -> rdev::Key {
-    use rdev::Key::*;
-    match &c {
-        'a' => KeyA,
-        '`' => BackQuote,
-        '1' => Num1,
-        '2' => Num2,
-        '3' => Num3,
-        '4' => Num4,
-        '5' => Num5,
-        '6' => Num6,
-        '7' => Num7,
-        '8' => Num8,
-        '9' => Num9,
-        '0' => Num0,
-        '-' => Minus,
-        '=' => Equal,
-        'q' => KeyQ,
-        'w' => KeyW,
-        'e' => KeyE,
-        'r' => KeyR,
-        't' => KeyT,
-        'y' => KeyY,
-        'u' => KeyU,
-        'i' => KeyI,
-        'o' => KeyO,
-        'p' => KeyP,
-        '[' => LeftBracket,
-        ']' => RightBracket,
-        's' => KeyS,
-        'd' => KeyD,
-        'f' => KeyF,
-        'g' => KeyG,
-        'h' => KeyH,
-        'j' => KeyJ,
-        'k' => KeyK,
-        'l' => KeyL,
-        ';' => SemiColon,
-        '\'' => Quote,
-        '\\' => BackSlash,
-        'z' => KeyZ,
-        'x' => KeyX,
-        'c' => KeyC,
-        'v' => KeyV,
-        'b' => KeyB,
-        'n' => KeyN,
-        'm' => KeyM,
-        ',' => Comma,
-        '.' => Dot,
-        '/' => Slash,
-        _ => Unknown(0),
-    }
-}
+/// Compute the minimal edit needed to transform what is currently displayed (`old`)
+/// into the desired output (`new`) by finding their longest common prefix.
+pub fn get_diff_parts<'a>(old: &str, new: &'a str) -> (usize, &'a str) {
+    let mut old_chars = old.chars();
+    let mut new_chars = new.char_indices();
 
-pub static mut KEYBOARD_LAYOUT_CHARACTER_MAP: OnceCell<HashMap<char, char>> = OnceCell::new();
-
-fn build_keyboard_layout_map(map: &mut HashMap<char, char>) {
-    map.clear();
-    let mut kb = Keyboard::new().unwrap();
-    for c in PREDEFINED_CHARS {
-        let key = rdev::EventType::KeyPress(get_key_from_char(c));
-        if let Some(s) = kb.add(&key) {
-            let ch = s.chars().last().unwrap();
-            map.insert(c, ch);
+    let mut common = 0usize;
+    let diverge_byte = loop {
+        match (old_chars.next(), new_chars.next()) {
+            (Some(a), Some((_, b))) if a == b => {
+                common += 1;
+            }
+            (_, Some((byte_pos, _))) => break byte_pos,
+            (_, None) => break new.len(),
         }
-    }
-}
+    };
 
-pub fn rebuild_keyboard_layout_map() {
-    unsafe {
-        if let Some(map) = KEYBOARD_LAYOUT_CHARACTER_MAP.get_mut() {
-            debug!("Rebuild keyboard layout map...");
-            build_keyboard_layout_map(map);
-            debug!("Done");
-        } else {
-            debug!("Creating keyboard layout map...");
-            let mut map = HashMap::new();
-            build_keyboard_layout_map(&mut map);
-            _ = KEYBOARD_LAYOUT_CHARACTER_MAP.set(map);
-            debug!("Done");
-        }
-    }
+    let old_len = old.chars().count();
+    let backspace_count = old_len.saturating_sub(common);
+    let suffix = &new[diverge_byte..];
+
+    (backspace_count, suffix)
 }
 
 #[allow(clippy::upper_case_acronyms)]
-#[derive(PartialEq, Eq, Data, Clone, Copy)]
+#[cfg_attr(feature = "druid", derive(druid::Data))]
+#[derive(PartialEq, Eq, Clone, Copy)]
 pub enum TypingMethod {
     VNI,
     Telex,
@@ -223,66 +140,6 @@ impl Display for TypingMethod {
             }
         )
     }
-}
-
-/// Compute the minimal edit needed to transform what is currently displayed (`old`)
-/// into the desired output (`new`) by finding their longest common prefix.
-///
-/// Returns `(backspace_count, suffix)` where:
-/// - `backspace_count` is the number of backspaces to send (to erase only the
-///   diverging tail of `old`)
-/// - `suffix` is the slice of `new` that must be typed after those backspaces
-///
-/// Both counts are in **Unicode scalar values** (chars), not bytes, because
-/// each backspace deletes one displayed character regardless of its byte width.
-/// The returned `suffix` is a byte slice of `new` starting at the first
-/// diverging char — no allocation, no `.collect()`.
-///
-/// # Example
-/// ```
-/// // old = "mô"  (on screen after typing "moo")
-/// // new = "mộ"  (engine output after pressing 'j' for nặng tone)
-/// // common prefix = "m"  → only "ô" needs deleting, only "ộ" needs typing
-/// let (bs, suffix) = get_diff_parts("mô", "mộ");
-/// assert_eq!(bs, 1);
-/// assert_eq!(suffix, "ộ");
-/// ```
-pub fn get_diff_parts<'a>(old: &str, new: &'a str) -> (usize, &'a str) {
-    // Walk both strings char-by-char simultaneously.
-    // We track the byte offset into `new` so we can return a zero-copy suffix slice.
-    let mut old_chars = old.chars();
-    let mut new_chars = new.char_indices();
-
-    // Number of chars that are identical from the start.
-    let mut common = 0usize;
-    // Byte offset in `new` where divergence begins (used for the suffix slice).
-    let mut diverge_byte = new.len(); // default: full match, empty suffix
-
-    loop {
-        match (old_chars.next(), new_chars.next()) {
-            (Some(a), Some((byte_pos, b))) if a == b => {
-                common += 1;
-                diverge_byte = byte_pos + b.len_utf8();
-            }
-            (_, Some((byte_pos, _))) => {
-                // Diverged — note byte position of the first differing char in `new`.
-                diverge_byte = byte_pos;
-                break;
-            }
-            (_, None) => {
-                // `new` is a prefix of (or equal to) `old` — no suffix to type.
-                diverge_byte = new.len();
-                break;
-            }
-        }
-    }
-
-    // old_tail_len = number of chars in old that are NOT part of the common prefix.
-    let old_len = old.chars().count();
-    let backspace_count = old_len.saturating_sub(common);
-    let suffix = &new[diverge_byte..];
-
-    (backspace_count, suffix)
 }
 
 pub struct InputState {
@@ -331,15 +188,12 @@ impl InputState {
         }
     }
 
-    pub fn update_active_app(&mut self) -> Option<()> {
-        let current_active_app = get_active_app_name();
-        // Only check if switch app
+    pub fn update_active_app(&mut self, current_active_app: &str) -> Option<()> {
         if current_active_app == self.active_app {
             return None;
         }
-        self.active_app = current_active_app;
+        self.active_app = current_active_app.to_string();
         let config = CONFIG_MANAGER.lock().unwrap();
-        // Only switch the input mode if we found the app in the config
         if config.is_vietnamese_app(&self.active_app) {
             self.enabled = true;
         }
@@ -347,6 +201,10 @@ impl InputState {
             self.enabled = false;
         }
         Some(())
+    }
+
+    pub fn active_app(&self) -> &str {
+        &self.active_app
     }
 
     pub fn set_temporary_disabled(&mut self) {
@@ -392,14 +250,10 @@ impl InputState {
         self.can_resume_previous_word = false;
     }
 
-    /// Mark that the previous word can be resumed if the user presses
-    /// backspace immediately (i.e. the word was ended by space/tab/enter).
     pub fn mark_resumable(&mut self) {
         self.can_resume_previous_word = true;
     }
 
-    /// Try to restore the previous word's buffers so editing can continue.
-    /// Returns true if the word was resumed, false otherwise.
     pub fn try_resume_previous_word(&mut self) -> bool {
         if !self.can_resume_previous_word || self.previous_word.is_empty() {
             return false;
@@ -415,11 +269,9 @@ impl InputState {
         if !self.is_macro_enabled {
             return None;
         }
-        // Exact match
         if let Some(target) = self.macro_table.get(&self.display_buffer) {
             return Some(target.clone());
         }
-        // Auto-capitalize: try lowercase lookup, then apply cap pattern
         if self.is_macro_autocap_enabled {
             let lower = self.display_buffer.to_lowercase();
             if let Some(target) = self.macro_table.get(&lower) {
@@ -501,9 +353,6 @@ impl InputState {
             .lock()
             .unwrap()
             .set_method(&method.to_string());
-        if let Some(event_sink) = UI_EVENT_SINK.get() {
-            _ = event_sink.submit_command(UPDATE_UI, (), Target::Auto);
-        }
     }
 
     pub fn get_method(&self) -> TypingMethod {
@@ -513,9 +362,6 @@ impl InputState {
     pub fn set_hotkey(&mut self, key_sequence: &str) {
         self.hotkey = Hotkey::from_str(key_sequence);
         CONFIG_MANAGER.lock().unwrap().set_hotkey(key_sequence);
-        if let Some(event_sink) = UI_EVENT_SINK.get() {
-            _ = event_sink.submit_command(UPDATE_UI, (), Target::Auto);
-        }
     }
 
     pub fn get_hotkey(&self) -> &Hotkey {
@@ -595,15 +441,7 @@ impl InputState {
         Ok(count)
     }
 
-    pub fn should_transform_keys(&self, c: &char) -> bool {
-        self.enabled
-    }
-
     pub fn transform_keys(&self) -> Result<(String, TransformResult), ()> {
-        // In w-literal mode (Telex only), replace standalone 'w' with a placeholder
-        // before feeding to the telex engine, then restore it in the output.
-        // A 'w' is considered standalone if NOT preceded by a Horn/Breve-eligible vowel
-        // (u, o for Horn; a for Breve). This preserves uw→ư, ow→ơ, aw→ă etc.
         let effective_buffer = if self.is_w_literal_enabled
             && matches!(self.method, TypingMethod::Telex | TypingMethod::TelexVNI)
         {
@@ -613,8 +451,6 @@ impl InputState {
         };
 
         if self.method == TypingMethod::TelexVNI {
-            // Try both methods; prefer VNI when the buffer contains digits
-            // (VNI's key differentiator), otherwise fall back to Telex.
             let buffer = effective_buffer;
             let result = std::panic::catch_unwind(move || {
                 let has_digits = buffer.chars().any(|c| c.is_ascii_digit());
@@ -624,7 +460,8 @@ impl InputState {
                     (output, transform_result)
                 } else {
                     let mut output = String::new();
-                    let transform_result = vi::telex::transform_buffer(buffer.chars(), &mut output);
+                    let transform_result =
+                        vi::telex::transform_buffer(buffer.chars(), &mut output);
                     let output = output.replace('\x01', "w").replace('\x02', "W");
                     (output, transform_result)
                 }
@@ -643,7 +480,6 @@ impl InputState {
                     vi::telex::transform_buffer(buffer.chars(), &mut output)
                 }
             };
-            // Restore masked standalone w's back to literal 'w'/'W'
             let output = if is_w_literal {
                 output.replace('\x01', "w").replace('\x02', "W")
             } else {
@@ -666,7 +502,7 @@ impl InputState {
         return DISMISS_APPS.iter().any(|app| self.active_app.contains(app));
     }
 
-    pub fn get_backspace_count(&self, is_delete: bool) -> usize {
+    pub fn get_backspace_count(&self, is_delete: bool, in_text_selection: bool) -> usize {
         let dp_len = self.display_buffer.chars().count();
         let backspace_count = if is_delete && dp_len >= 1 {
             dp_len
@@ -674,7 +510,7 @@ impl InputState {
             dp_len - 1
         };
 
-        if is_in_text_selection() {
+        if in_text_selection {
             backspace_count + 1
         } else {
             backspace_count
@@ -769,23 +605,42 @@ impl InputState {
         let config = CONFIG_MANAGER.lock().unwrap();
         return config.is_allowed_word(word);
     }
+
+    pub fn should_restore_word(&self) -> bool {
+        let typing_buffer = self.get_typing_buffer();
+        let display_buffer = self.get_displaying_word();
+
+        let is_transformed_word = typing_buffer != display_buffer;
+        if !is_transformed_word {
+            return false;
+        }
+
+        let is_valid_word = vi::validation::is_valid_word(display_buffer);
+        if is_valid_word {
+            return false;
+        }
+
+        if self.is_allowed_word(display_buffer) {
+            return false;
+        }
+
+        let is_vni_numeric_shortcut =
+            self.method == TypingMethod::VNI && typing_buffer.chars().any(|c| c.is_numeric());
+        !is_vni_numeric_shortcut
+    }
 }
 
 #[cfg(test)]
 mod diff_tests {
     use super::get_diff_parts;
 
-    // ── Basic tone application ────────────────────────────────────────────────
-
-    /// "mô" → "mộ": only the vowel+tone char is replaced, "m" stays.
     #[test]
     fn tone_on_vowel_preserves_consonant_prefix() {
         let (bs, sfx) = get_diff_parts("mô", "mộ");
-        assert_eq!(bs, 1, "should delete only 'ô'");
+        assert_eq!(bs, 1);
         assert_eq!(sfx, "ộ");
     }
 
-    /// "mo" → "mô": typing 'o' again applies the circumflex.
     #[test]
     fn circumflex_application() {
         let (bs, sfx) = get_diff_parts("mo", "mô");
@@ -793,33 +648,26 @@ mod diff_tests {
         assert_eq!(sfx, "ô");
     }
 
-    /// "tieng" → "tiếng": "ti" preserved, vowel+tone suffix replaced.
     #[test]
     fn multi_char_prefix_preserved() {
         let (bs, sfx) = get_diff_parts("tieng", "tiếng");
-        assert_eq!(bs, 3); // "eng" deleted
+        assert_eq!(bs, 3);
         assert_eq!(sfx, "ếng");
     }
 
-    /// "nguyen" → "nguyên": "nguy" is common.
     #[test]
     fn longer_common_prefix() {
         let (bs, sfx) = get_diff_parts("nguyen", "nguyên");
-        assert_eq!(bs, 2); // "en" deleted
+        assert_eq!(bs, 2);
         assert_eq!(sfx, "ên");
     }
 
-    // ── No-op / identical strings ─────────────────────────────────────────────
-
-    /// Identical strings → 0 backspaces, empty suffix.
     #[test]
     fn identical_strings_no_op() {
         let (bs, sfx) = get_diff_parts("mộ", "mộ");
         assert_eq!(bs, 0);
         assert_eq!(sfx, "");
     }
-
-    // ── Empty edge cases ──────────────────────────────────────────────────────
 
     #[test]
     fn both_empty() {
@@ -842,25 +690,19 @@ mod diff_tests {
         assert_eq!(sfx, "");
     }
 
-    // ── Prefix / suffix relationships ─────────────────────────────────────────
-
-    /// new is a strict prefix of old: delete tail, type nothing.
     #[test]
     fn new_is_prefix_of_old() {
         let (bs, sfx) = get_diff_parts("mộng", "mộ");
-        assert_eq!(bs, 2); // delete "ng"
+        assert_eq!(bs, 2);
         assert_eq!(sfx, "");
     }
 
-    /// old is a strict prefix of new: 0 backspaces, append tail.
     #[test]
     fn old_is_prefix_of_new() {
         let (bs, sfx) = get_diff_parts("mộ", "mộng");
         assert_eq!(bs, 0);
         assert_eq!(sfx, "ng");
     }
-
-    // ── Completely different strings ──────────────────────────────────────────
 
     #[test]
     fn no_common_prefix() {
@@ -869,14 +711,10 @@ mod diff_tests {
         assert_eq!(sfx, "xyz");
     }
 
-    // ── Multi-byte / Unicode correctness ─────────────────────────────────────
-
-    /// Each Vietnamese toned vowel is 1 char, possibly 3 bytes.
-    /// backspace_count must be in chars, not bytes.
     #[test]
     fn char_count_not_byte_count() {
         let (bs, sfx) = get_diff_parts("ộ", "ô");
-        assert_eq!(bs, 1, "one char deleted, not three bytes");
+        assert_eq!(bs, 1);
         assert_eq!(sfx, "ô");
     }
 
@@ -887,9 +725,6 @@ mod diff_tests {
         assert_eq!(sfx, "ể");
     }
 
-    // ── Realistic Telex sequences ─────────────────────────────────────────────
-
-    /// "moo" (buffer) → "mô" (engine output).
     #[test]
     fn telex_moo_to_mo_hat() {
         let (bs, sfx) = get_diff_parts("moo", "mô");
@@ -897,7 +732,6 @@ mod diff_tests {
         assert_eq!(sfx, "ô");
     }
 
-    /// "cas" → "cá": "c" preserved.
     #[test]
     fn telex_cas_to_ca_sac() {
         let (bs, sfx) = get_diff_parts("cas", "cá");
@@ -905,23 +739,19 @@ mod diff_tests {
         assert_eq!(sfx, "á");
     }
 
-    /// "viet" → "việt"
     #[test]
     fn telex_viet_transform() {
         let (bs, sfx) = get_diff_parts("viet", "việt");
-        assert_eq!(bs, 2); // common = "vi"
+        assert_eq!(bs, 2);
         assert_eq!(sfx, "ệt");
     }
 
-    /// Tone cycling: "tiến" → "tiền" (sắc → huyền), "ti" preserved.
     #[test]
     fn tone_cycling_preserves_prefix() {
         let (bs, sfx) = get_diff_parts("tiến", "tiền");
         assert_eq!(bs, 2);
         assert_eq!(sfx, "ền");
     }
-
-    // ── Suffix slice is a zero-copy view into `new` ───────────────────────────
 
     #[test]
     fn suffix_is_valid_utf8_slice_of_new() {
@@ -941,7 +771,6 @@ mod mask_w_tests {
 
     #[test]
     fn standalone_w_is_masked() {
-        // 'w' not preceded by eligible vowel → masked
         assert_eq!(mask_standalone_w("w"), "\x01");
         assert_eq!(mask_standalone_w("rw"), "r\x01");
     }
@@ -954,7 +783,6 @@ mod mask_w_tests {
 
     #[test]
     fn w_after_eligible_vowel_is_not_masked() {
-        // aw→ă, uw→ư, ow→ơ should pass through
         assert_eq!(mask_standalone_w("aw"), "aw");
         assert_eq!(mask_standalone_w("uw"), "uw");
         assert_eq!(mask_standalone_w("ow"), "ow");
@@ -962,7 +790,6 @@ mod mask_w_tests {
 
     #[test]
     fn ww_after_eligible_vowel_not_masked() {
-        // "aww" → both w's passed through so telex sees "ww" and undoes breve
         assert_eq!(mask_standalone_w("aww"), "aww");
         assert_eq!(mask_standalone_w("uww"), "uww");
         assert_eq!(mask_standalone_w("oww"), "oww");
@@ -971,7 +798,6 @@ mod mask_w_tests {
 
     #[test]
     fn standalone_ww_both_masked() {
-        // "ww" with no eligible vowel before → both masked
         assert_eq!(mask_standalone_w("ww"), "\x01\x01");
         assert_eq!(mask_standalone_w("rww"), "r\x01\x01");
     }
@@ -1009,22 +835,18 @@ mod tracking_tests {
 
     #[test]
     fn pop_to_empty_then_new_word_re_enables_tracking() {
-        // Simulates: type "raww" → stop_tracking → backspace to empty → new_word
         let mut state = InputState::new();
         state.push('r');
         state.push('a');
         state.push('w');
         state.push('w');
-        state.stop_tracking(); // triggered by "ww" pattern
+        state.stop_tracking();
         assert!(!state.is_tracking());
         assert!(state.is_buffer_empty());
 
-        // Backspaces clear the screen (handled by OS), buffer already empty.
-        // Calling new_word() re-enables tracking for the next keystrokes.
         state.new_word();
         assert!(state.is_tracking());
 
-        // New characters should be tracked
         state.push('o');
         state.push('o');
         assert_eq!(state.get_typing_buffer(), "oo");
@@ -1037,12 +859,10 @@ mod tracking_tests {
         state.push('e');
         state.push('s');
         state.push('t');
-        // Simulate end-of-word (space) → new_word + mark_resumable
         state.new_word();
         state.mark_resumable();
         assert!(state.is_buffer_empty());
 
-        // Resume should restore the previous word
         assert!(state.try_resume_previous_word());
         assert!(state.is_tracking());
         assert_eq!(state.get_typing_buffer(), "test");
