+++ core/src/keycode.rs
@@ -1,0 +1,5 @@
+pub const KEY_ENTER: char = '\x13';
+pub const KEY_SPACE: char = '\u{0020}';
+pub const KEY_TAB: char = '\x09';
+pub const KEY_DELETE: char = '\x08';
+pub const KEY_ESCAPE: char = '\x26';
